Improve styling of sidebar widget sections
Modify the app.js and app.css classes to apply Tailwind directives to
widgets except for the search bar and login/logout. Add to the app.js
file so that it prepends an empty div to the sections that can be used
for the skewed rounded rectangle Tailwind background. Also apply the
class stateless-sage-sidebar-top-level to each of the widgets to allow
for applying directives to the top-level but none of the nested
wp-block-group elements.